+++ components/modal/addToFolderModal/AddToFolderModal.tsx
@@ -1,14 +1,13 @@
-import styled from "styled-components";
-import { modalState } from "../../recoil/modal";
-import { useRecoilValue, useResetRecoilState } from "recoil";
-
-import { useEffect, useState } from "react";
+import { modalState } from "../../../recoil/modal";
+import { useResetRecoilState } from "recoil";
+import { useState } from "react";
 import CloseIcon from "@/public/assets/modal/img_closeIcon.png";
 import CheckIcon from "@/public/assets/modal/img_checkIcon.png";
 import Image from "next/image";
-import DefaultBtn from "../button/Button";
-import { ModalMainContainer } from "./ModalStyledComponents";
+import DefaultBtn from "../../button/Button";
+import { ModalMainContainer } from "../ModalStyledComponents";
 import { FolderData, SelectedFolder } from "@/pages/folder";
+import { FolderContainer } from "./addToFolderModalStyled";
 
 interface AddToFolderModalProp {
   folderData: FolderData[];
@@ -71,33 +70,3 @@ const AddToFolderModal = ({
 };
 
 export default AddToFolderModal;
-
-const FolderContainer = styled.div<{ $active: boolean }>`
-  display: flex;
-  gap: 0.8rem;
-  padding: 0.8rem;
-  width: 100%;
-  box-sizing: border-box;
-  cursor: pointer;
-  align-items: center;
-  border-radius: ${(props) => (props.$active ? "0.8rem" : "0")};
-  background: ${(props) => (props.$active ? "#f0f6ff" : "#fff")};
-
-  .title {
-    color: ${(props) => (props.$active ? "#6D6AFE" : "var(--gray100)")};
-    font-size: 1.6rem;
-    font-weight: 400;
-    line-height: 2.4rem;
-  }
-
-  .linkNum {
-    color: var(--gray60);
-    font-size: 1.4rem;
-    font-weight: 400;
-  }
-
-  .checkIcon {
-    margin-left: auto;
-    display: ${(props) => (props.$active ? "flex" : "none")};
-  }
-`;
