+++ components/modal/defaultModalLayout/DefaultModalLayout.tsx
@@ -1,12 +1,9 @@
-import styled from "styled-components";
 import { useResetRecoilState } from "recoil";
-import { ModalMainContainer } from "./ModalStyledComponents";
-import { modalState } from "../../recoil/modal";
+import { modalState } from "../../../recoil/modal";
 import CloseIcon from "@/public/assets/modal/img_closeIcon.png";
 import { ReactNode } from "react";
 import Image from "next/image";
-import { ButtonContainer } from "../button/Button";
-import { InputContainer } from "../input/Input";
+import { ModalMain } from "./defaultModalLayoutStyled";
 
 interface DefaultModalProp {
   title: string;
@@ -47,19 +44,3 @@ const DefaultModalLayout = ({
 };
 
 export default DefaultModalLayout;
-
-const ModalMain = styled(ModalMainContainer)<{ $isDelete: boolean }>`
-  ${InputContainer} {
-    width: 100%;
-    background: var(--white);
-    border: 1px solid var(--gray20);
-  }
-
-  ${ButtonContainer} {
-    margin-top: ${(props) => (props.$isDelete ? "0" : "-0.8rem")};
-  }
-
-  ${InputContainer}:focus-within {
-    border: 1px solid var(--linkbrary-primary-color, #6d6afe);
-  }
-`;
